Extract SetupJavaDistributionReplacerVisitor to new file
Rename recipes for consistency
